requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
azure-functions==1.11.2
    # via -r requirements.in
blis==0.9.1
    # via thinc
catalogue==2.0.8
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2022.9.14
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via typer
confection==0.0.1
    # via thinc
cymem==2.0.6
    # via
    #   preshed
    #   spacy
    #   thinc
idna==3.4
    # via requests
jinja2==3.1.2
    # via spacy
langcodes==3.3.0
    # via spacy
markupsafe==2.1.1
    # via jinja2
murmurhash==1.0.8
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==1.23.3
    # via
    #   spacy
    #   thinc
oauthlib==3.2.1
    # via
    #   requests-oauthlib
    #   tweepy
packaging==21.3
    # via spacy
pathy==0.6.2
    # via spacy
pillow==9.2.0
    # via -r requirements.in
preshed==3.0.7
    # via
    #   spacy
    #   thinc
pydantic==1.9.2
    # via
    #   spacy
    #   thinc
pyparsing==3.0.9
    # via packaging
requests==2.28.1
    # via
    #   requests-oauthlib
    #   spacy
    #   tweepy
requests-oauthlib==1.3.1
    # via tweepy
smart-open==5.2.1
    # via pathy
spacy==3.4.1
    # via -r requirements.in
spacy-legacy==3.0.10
    # via spacy
spacy-loggers==1.0.3
    # via spacy
srsly==2.4.4
    # via
    #   spacy
    #   thinc
thinc==8.1.1
    # via spacy
tqdm==4.64.1
    # via spacy
tweepy==4.10.1
    # via -r requirements.in
typer==0.4.2
    # via
    #   pathy
    #   spacy
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.12
    # via requests
wasabi==0.10.1
    # via
    #   spacy
    #   spacy-loggers
    #   thinc

# The following packages are considered to be unsafe in a requirements file:
# setuptools
